+++ conflate/conflate.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 import argparse
 import codecs
+import json
 import kdtree
 import logging
 import math
@@ -8,10 +9,6 @@ import requests
 import os
 import sys
 from io import BytesIO
-import json    # for profiles
-import re      # for profiles
-import zipfile # for profiles
-from collections import defaultdict # for profiles
 try:
     from .version import __version__
 except ImportError:
@@ -81,6 +78,7 @@ class OSMPoint(SourcePoint):
         self.version = version
         self.members = None
         self.action = None
+        self.categories = categories or set()
         self.remarks = None
 
     def copy(self):
@@ -210,27 +208,38 @@ class OsmConflator:
         (k, v) turns into [k=v], (k,) into [k], (k, None) into [!k], (k, "~v") into [k~v]."""
         tags = self.profile.get(
             'query', required="a list of tuples. E.g. [('amenity', 'cafe'), ('name', '~Mc.*lds')]")
+        tag_strs = []
         if isinstance(tags, str):
-            tag_str = tags
+            tag_strs = [tags]
         else:
-            tag_str = ''
-            for t in tags:
-                if len(t) == 1:
-                    q = '"{}"'.format(t[0])
-                elif t[1] is None or len(t[1]) == 0:
-                    q = '"!{}"'.format(t[0])
-                elif t[1][0] == '~':
-                    q = '"{}"~"{}"'.format(t[0], t[1][1:])
-                else:
-                    q = '"{}"="{}"'.format(t[0], t[1])
-                tag_str += '[' + q + ']'
+            if not isinstance(tags[0], str) and isinstance(tags[0][0], str):
+                tags = [tags]
+            for tags_q in tags:
+                if isinstance(tags_q, str):
+                    tag_strs.append(tags_q)
+                    continue
+                tag_str = ''
+                for t in tags_q:
+                    if len(t) == 1:
+                        q = '"{}"'.format(t[0])
+                    elif t[1] is None or len(t[1]) == 0:
+                        q = '"!{}"'.format(t[0])
+                    elif t[1][0] == '~':
+                        q = '"{}"~"{}",i'.format(t[0], t[1][1:])
+                    elif len(t) > 2:
+                        q = '"{}"~"^({})$"'.format(t[0], '|'.join(t[1:]))
+                    else:
+                        q = '"{}"="{}"'.format(t[0], t[1])
+                    tag_str += '[' + q + ']'
+                tag_strs.append(tag_str)
 
         timeout = self.profile.get('overpass_timeout', 120)
         query = '[out:xml]{};('.format('' if timeout is None else '[timeout:{}]'.format(timeout))
         for bbox in bboxes:
             bbox_str = '' if bbox is None else '(' + ','.join([str(x) for x in bbox]) + ')'
-            for t in ('node', 'way', 'relation["type"="multipolygon"]'):
-                query += t + tag_str + bbox_str + ';'
+            for tag_str in tag_strs:
+                for t in ('node', 'way', 'relation["type"="multipolygon"]'):
+                    query += t + tag_str + bbox_str + ';'
         if self.ref is not None:
             for t in ('node', 'way', 'relation'):
                 query += t + '["' + self.ref + '"];'
@@ -349,21 +358,68 @@ class OsmConflator:
         padding = self.profile.get('bbox_padding', BBOX_PADDING)
         return [get_bbox(b, padding) for b in boxes]
 
-    def check_against_profile_tags(self, tags):
+    def get_categories(self, tags):
+        def match_query(tags, query):
+            for tag in query:
+                if len(tag) == 1:
+                    if tag[0] in tags:
+                        return False
+                    elif tag[1] is None or tag[1] == '':
+                        if tag[0] not in tags:
+                            return False
+                    else:
+                        value = tags.get(tag[0], None)
+                        if value is None:
+                            return False
+                        found = False
+                        for t2 in tag[1:]:
+                            if t2[0] == '~':
+                                m = re.search(t2[1:], value)
+                                if not m:
+                                    return False
+                            elif t2[0] == '!':
+                                if t2[1:].lower() in value.lower():
+                                    found = True
+                            elif t2 == value:
+                                found = True
+                            if found:
+                                break
+                        if not found:
+                            return False
+            return True
+
+        def tags_to_query(tags):
+            return [(k, v) for k, v in tags.items()]
+
+        result = set()
         qualifies = self.profile.get('qualifies', args=tags)
         if qualifies is not None:
-            return qualifies
+            if qualifies:
+                result.add(None)
+            return result
 
+        # First check default query
         query = self.profile.get('query', None)
-        if query is not None and not isinstance(query, str):
-            for tag in query:
-                if len(tag) >= 1:
-                    if tag[0] not in tags:
-                        return False
-                    if len(tag) >= 2 and tag[1][0] != '~':
-                        if tag[1] != tags[tag[0]]:
-                            return False
-        return True
+        if query is not None:
+            if isinstance(query, str):
+                result.add(None)
+            else:
+                if isinstance(query[0][0], str):
+                    query = [query]
+                for q in query:
+                    if match_query(tags, q):
+                        result.add(None)
+                        break
+
+        # Then check each category if we got these
+        categories = self.profile.get('categories', {})
+        for name, params in categories.items():
+            if 'tags' not in params and 'query' not in params:
+                raise ValueError('No tags and query attributes for category "{}"'.format(name))
+            if match_query(tags, params.get('query', tags_to_query(params.get('tags')))):
+                result.add(name)
+
+        return result
 
     def download_osm(self):
         """Constructs an Overpass API query and requests objects
@@ -424,7 +480,8 @@ class OsmConflator:
             tags = {}
             for tag in el.findall('tag'):
                 tags[tag.get('k')] = tag.get('v')
-            if not self.check_against_profile_tags(tags):
+            categories = self.get_categories(tags)
+            if categories is False or categories is None or len(categories) == 0:
                 continue
 
             if el.tag == 'node':
@@ -458,7 +515,7 @@ class OsmConflator:
                 continue
             pt = OSMPoint(
                 el.tag, int(el.get('id')), int(el.get('version')),
-                coord[0], coord[1], tags)
+                coord[0], coord[1], tags, categories)
             pt.members = members
             if pt.is_poi():
                 if callable(weight_fn):
@@ -649,7 +706,8 @@ class OsmConflator:
                 nearest = [p for p in nearest if match_func(p[0].data.tags, point.tags)]
                 if not nearest:
                     return None, None
-            nearest = [(n[0], n[0].data.distance(point)) for n in nearest]
+            nearest = [(n[0], n[0].data.distance(point))
+                       for n in nearest if point.category in n[0].data.categories]
             return sorted(nearest, key=lambda kv: kv[1])[0]
 
         if not self.osmdata:
@@ -849,6 +907,22 @@ def read_dataset(profile, fileobj):
         required='returns a list of SourcePoints with the dataset')
 
 
+def add_categories_to_dataset(profile, dataset):
+    categories = profile.get('categories')
+    if not categories:
+        return
+    tag = profile.get('category_tag')
+    other = categories.get('other', {})
+    for d in dataset:
+        if tag and tag in d.tags:
+            d.category = d.tags[tag]
+            del d.tags[tag]
+        if d.category:
+            cat_tags = categories.get(d.category, other).get('tags', None)
+            if cat_tags:
+                d.tags.update(cat_tags)
+
+
 def transform_dataset(profile, dataset):
     """Transforms tags in the dataset using the "transform" method in the profile
     or the instructions in that field in string or dict form."""
@@ -918,6 +992,56 @@ def transform_dataset(profile, dataset):
             d.tags[key] = value
 
 
+def write_for_filter(profile, dataset, f):
+    def query_to_tag_strings(query):
+        if isinstance(query, str):
+            raise ValueError('Query string for filter should not be a string')
+        result = []
+        if not isinstance(query[0], str) and isinstance(query[0][0], str):
+            query = [query]
+        for q in query:
+            if isinstance(q, str):
+                raise ValueError('Query string for filter should not be a string')
+            parts = []
+            for part in q:
+                if len(part) == 1:
+                    parts.append(part[0])
+                elif part[1] is None or len(part[1]) == 0:
+                    parts.append('{}='.format(part[0]))
+                elif part[1][0] == '~':
+                    raise ValueError('Cannot use regular expressions in filter')
+                elif '|' in part[1] or ';' in part[1]:
+                    raise ValueError('"|" and ";" symbols is not allowed in query values')
+                else:
+                    parts.append('='.join(part))
+            result.append('|'.join(parts))
+        return result
+
+    def tags_to_query(tags):
+        return [(k, v) for k, v in tags.items()]
+
+    categories = profile.get('categories', {})
+    p_query = profile.get('query', None)
+    if p_query is not None:
+        categories[None] = {'query': p_query}
+    cat_map = {}
+    i = 0
+    try:
+        for name, query in categories.items():
+            for tags in query_to_tag_strings(query.get('query', tags_to_query(query.get('tags')))):
+                f.write('{},{},{}\n'.format(i, name or '', tags))
+            cat_map[name] = i
+            i += 1
+    except ValueError as e:
+        logging.error(e)
+        return False
+    f.write('\n')
+    for d in dataset:
+        if d.category in cat_map:
+            f.write('{},{},{}\n'.format(d.lon, d.lat, cat_map[d.category]))
+    return True
+
+
 def run(profile=None):
     parser = argparse.ArgumentParser(
         description='''{}.
@@ -928,15 +1052,17 @@ def run(profile=None):
     parser.add_argument('-i', '--source', type=argparse.FileType('rb'), help='Source file to pass to the profile dataset() function')
     parser.add_argument('-a', '--audit', type=argparse.FileType('r'), help='Conflation validation result as a JSON file')
     parser.add_argument('-o', '--output', type=argparse.FileType('w'), help='Output OSM XML file name')
+    parser.add_argument('-p', '--param', help='Optional parameter for the profile')
     parser.add_argument('--osc', action='store_true', help='Produce an osmChange file instead of JOSM XML')
     parser.add_argument('--osm', help='Instead of querying Overpass API, use this unpacked osm file. Create one from Overpass data if not found')
     parser.add_argument('-c', '--changes', type=argparse.FileType('w'), help='Write changes as GeoJSON for visualization')
     parser.add_argument('-m', '--check-move', action='store_true', help='Check for moveability of modified modes')
+    parser.add_argument('-f', '--for-filter', type=argparse.FileType('w'), help='Prepare a file for the filtering script')
     parser.add_argument('--verbose', '-v', action='store_true', help='Display debug messages')
     parser.add_argument('--quiet', '-q', action='store_true', help='Do not display informational messages')
     options = parser.parse_args()
 
-    if not options.output and not options.changes:
+    if not options.output and not options.changes and not options.for_filter:
         parser.print_help()
         return
 
@@ -952,6 +1078,8 @@ def run(profile=None):
 
     if not profile:
         logging.debug('Loading profile %s', options.profile)
+    global param
+    param = options.param
     profile = Profile(profile or options.profile)
 
     dataset = read_dataset(profile, options.source)
@@ -959,7 +1087,13 @@ def run(profile=None):
         logging.error('Empty source dataset')
         sys.exit(2)
     transform_dataset(profile, dataset)
+    add_categories_to_dataset(profile, dataset)
     logging.info('Read %s items from the dataset', len(dataset))
+
+    if options.for_filter:
+        if write_for_filter(profile, dataset, options.for_filter):
+            logging.info('Prepared data for filtering, exitting')
+        return
 
     audit = None
     if options.audit: